+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2026-05-25
 
 - Put a cap on message length
+- Changed the passkey log-in flow, use your browser's autofill/dropdown instead.
 
 ## 2026-05-24